MHV-58512: Unit tests written for download all records helper functions
