+++ src/applications/mhv-medical-records/tests/containers/LandingPage.unit.spec.jsx
@@ -43,6 +43,25 @@ describe('Landing Page', () => {
     expect(screen).to.exist;
   });
 
+  it('displays a section linking to My HealtheVet classic to download all records', () => {
+    const screen = renderWithStoreAndRouter(<LandingPage />, {});
+    expect(
+      screen.getByText('Download your VA medical records', {
+        selector: 'h2',
+        exact: true,
+      }),
+    ).to.exist;
+    expect(
+      screen.getAllByText(
+        'Go to medical records on the My HealtheVet website',
+        {
+          selector: 'a',
+          exact: true,
+        },
+      ).length,
+    ).to.eq(2);
+  });
+
   it('displays downtimeNotification when downtimeApproaching is true', () => {
     const customState = {
       featureToggles: {},
